Commented out broken code to push

+++ classes/GameObject/GameObject.h
@@ -18,7 +18,7 @@ const GLfloat colours[7][3] = {
        { 0.0, 0.0, 0.6 }  // dark-blue
 };
 
-const int Gravity = 1;
+const float Gravity = 1;
 
 class GameObject {
   public:
